requirements.txt
openai
datasets
pandas
python-dotenv
tqdm
numpy
nemoguardrails
langchain==0.2.17
langchain-chroma==0.1.4
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.143